Call prettyprint only once (#4850)
