+++ src/actions/sources/createPrettySource.js
@@ -2,9 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at <http://mozilla.org/MPL/2.0/>. */
 
-import { getFrames, getSource } from "../../selectors";
+import { getSource } from "../../selectors";
 import { prettyPrint } from "../../workers/pretty-print";
-import { updateFrameLocations } from "../../utils/pause";
 import { getPrettySourceURL } from "../../utils/source";
 
 export function createPrettySource(sourceId) {
@@ -13,31 +12,21 @@ export function createPrettySource(sourceId) {
     const url = getPrettySourceURL(source.get("url"));
     const id = await sourceMaps.generatedToOriginalId(sourceId, url);
 
-    const { code, mappings } = await prettyPrint({
-      source,
-      url
-    });
-
-    await sourceMaps.applySourceMap(source.get("id"), url, code, mappings);
-
-    let frames = getFrames(getState());
-    if (frames) {
-      frames = await updateFrameLocations(frames, sourceMaps);
-    }
-
     const prettySource = {
       url,
       id,
       isPrettyPrinted: true,
-      text: code,
       contentType: "text/javascript",
-      frames,
-      loadedState: "loaded"
+      loadedState: "loading"
     };
+    dispatch({ type: "ADD_SOURCE", source: prettySource });
+
+    const { code, mappings } = await prettyPrint({ source, url });
+    await sourceMaps.applySourceMap(source.get("id"), url, code, mappings);
 
     dispatch({
-      type: "ADD_SOURCE",
-      source: prettySource
+      type: "UPDATE_SOURCE",
+      source: { ...prettySource, text: code, loadedState: "loaded" }
     });
 
     return prettySource;
