+++ src/reducers/sources.js
@@ -54,6 +54,11 @@ function update(
   let location = null;
 
   switch (action.type) {
+    case "UPDATE_SOURCE": {
+      const source = action.source;
+      return updateSource(state, source);
+    }
+
     case "ADD_SOURCE": {
       return updateSource(state, action.source);
     }
